Stats follow-ups: drop basics from top cards, mana-symbol colors, card-type legend, public profiles
Addresses review feedback:

- Exclude basic lands from "Most-played cards" — they're in every deck and
  drowned out the cards that actually characterize a player. Filtered via the
  registry's `isBasicLand` after the group-by.
- "Colors you play" now renders mana-font pip symbols + a proportional bar +
  count per identity, instead of an axis-labelled bar chart.
- "Card types" pie gains a labelled legend (type name + count + swatch), so it's
  clear which slice is which.
- View other players' profiles: new public, read-only `GET /api/stats/users/{id}`
  bundling a player's whole dashboard in one response (404 if unknown; decklists
  stay owner-only). New `/u/:userId` page renders the shared `StatsDashboard`
  plus a read-only recent-games list. Head-to-head opponents with an account and
  the friends list now link to these profiles.

Refactor: extracted the stats rendering from StatsPage into a shared
`StatsDashboard` so /stats and /u/:userId stay thin data-fetchers and the three
display improvements above land in both. New controller helpers
`statsFor`/`ratingsFor`/`ratingHistoryFor` back both `/me/*` and the bundle.
